refactor: 컴포넌트 로드 기능을 별도의 파일로 분리
- IncludeComponent.js 파일에 있었던 컴포넌트 로드 스크립트를 ComponentCheckFunction.js 파일로 분리
- 코드 재사용성 향상
- JS 파일 이름 수정
