+++ index.html
@@ -16,7 +16,7 @@
 
   <div data-include-html="/src/pages/home.html"></div>
 
-  <script src="/src/js/IncludeHtml.js"></script>
+  <script type="module" src="/src/js/HtmlLoader.js"></script>
 </body>
 
 </html>
