+++ src/pages/privacy/index.html
@@ -15,7 +15,7 @@
 
   <div data-include-component="/src/components/privacy/Privacy.html"></div>
 
-  <script src="/src/js/IncludeComponent.js"></script>
+  <script type="module" src="/src/js/ComponentLoader.js"></script>
 </body>
 
 </html>
