+++ src/pages/signup/index.html
@@ -20,7 +20,7 @@
 
   <div data-include-component="/src/components/signup/Signup.html"></div>
 
-  <script src="/src/js/IncludeComponent.js"></script>
+  <script type="module" src="/src/js/ComponentLoader.js"></script>
 </body>
 
 </html>
